// src/types/colors.ts
export interface ColorScheme {
  primary: string;
  secondary: string;
  accent: string;
  background: string;
  heroBackground: string;
  text: string;
  headlineColor: string;
  subheadlineColor: string;
  featuresColor: string;
  featuresTextColor: string;
  aboutColor: string;
  aboutTextColor: string;
  contactColor: string;
  contactTextColor: string;
}